Move CreateBook to new file addbook.
